+++ src/Endpoint/MyContactsEndpoint.php
@@ -4,6 +4,10 @@ declare(strict_types=1);
 
 namespace DigitalCz\DigiSign\Endpoint;
 
+use DigitalCz\DigiSign\Endpoint\Traits\CreateEndpointTrait;
+use DigitalCz\DigiSign\Endpoint\Traits\DeleteEndpointTrait;
+use DigitalCz\DigiSign\Endpoint\Traits\ListEndpointTrait;
+use DigitalCz\DigiSign\Endpoint\Traits\UpdateEndpointTrait;
 use DigitalCz\DigiSign\Resource\Collection;
 use DigitalCz\DigiSign\Resource\Contact;
 
@@ -12,6 +16,11 @@ use DigitalCz\DigiSign\Resource\Contact;
  */
 final class MyContactsEndpoint extends ResourceEndpoint
 {
+    use CreateEndpointTrait;
+    use DeleteEndpointTrait;
+    use ListEndpointTrait;
+    use UpdateEndpointTrait;
+
     public function __construct(MyEndpoint $parent)
     {
         parent::__construct($parent, '/contacts', Contact::class);